requirements.txt
bootstrap-admin==0.4.3
Django==2.1.5
django-rest-auth==0.9.5
djangorestframework==3.9.4
gitdb2==2.0.5
GitPython==2.1.11
Pygments==2.4.2
python-git==2018.2.1
pytz==2019.1
Send2Trash==1.5.0
six==1.12.0
smmap2==2.0.5
sqlparse==0.3.0
